// src/app/oapdis/teacher/model/teacher.ts
export class Teacher {
    id: string;
    userId: string;
    firstName: string;
    lastName: string;
    task: string;
    title: string;
    email: string;
    password: string;
    phoneNumber: string;
    departmentId: number;
    typeId: number;

    projectCount: number;
    departmentName: string;
}